add full israel adapter
this israel returns an object for each region with measurements from each station in openAQ format.
One thing it currently lacks is info about the interval of each adapter. I am holding off here to see if
the work @dolugen is doing currently for South Africa can be implemented here in israel.
upon figuring that out we'll add this last section accordingly.