Hotfix: remove `results` subkey from task results
This fixes the underlying cause of openaq/openaq-api#351 which was
introduced by openaq/openaq-fetch#354

Each fetch task generated a message that contained information about
that task. To migrate to S3, I attached measurement results to these
messages. However I forgot to remove the property before pushing these
fetches to the PG database. This results in a bloated fetches table that
affects the API availibility for the fetch endpoint.