Adds usage of streams into fetch.js and eee-direct adapter.
Fetch now allows two types of adapters: legacy array ones and those returning a `scramjet` stream of items. In the first case the `fetchData` method is called as before. In the
latter case a new `fetchStream` method will be called. In case two methods are exposed `fetchStream` is preferred as the more optimal one.B
All entries are verified and parsed into the DB format using stream transforms.